Switch to using pydantic instead of jsonschema
Pydantic allows for the same validation, but with the schema written in python code. And, typing info will be correct throughout, and can use class attributes style instead of a dict with string keys for using the from-yaml info. Also, when needed can still generate a jsonschema, for external tools to check the yaml that we want.